requirements.txt
yt-dlp
openai-whisper
ffmpeg-python
requests
python-dotenv
certifi